fix(animal): corrigido tipo de dado para o atributo de tamanho

+++ api/src/main/java/com/ong/acolhepatinhas/api/animal/DTO/EditAnimalRequest.java
@@ -1,5 +1,6 @@
 package com.ong.acolhepatinhas.api.animal.DTO;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.LocalDate;
 
@@ -58,7 +59,7 @@ public record EditAnimalRequest(
     Integer size,
 
     @Schema(example = "1.5", description = "Valor em kg")
-    Double weight,
+    BigDecimal weight,
 
     boolean neutered
 ) {
